Keplerian orbits (bounded) round trip correctly (I think correctly, but they round trip for sure)

+++ HumanSpaceProgram/Assets/HSP.CelestialBodies/CelestialBodyFactory.cs
@@ -18,8 +18,8 @@ namespace HSP.CelestialBodies
 
         //const float radius = 1000; //6371000f; // m
         //const float mass = 20e16f; //5.97e24f; // kg  // 20e16f for 1km radius is good
-        public float radius = 6371000f;
-        public float mass = 5.97e24f;
+        public double radius = 6371000.0;
+        public double mass = 5.97e24;
         public const int subdivs = 7; // 7 is the maximum value for a single plane that won't cause issues here.
 
         public CelestialBodyFactory( string id )
